+++ MarcoSmiles/Assets/Own/Scripts/_GM.cs
@@ -33,7 +33,7 @@ public class _GM : MonoBehaviour
 
 
     public static bool isActive = false;                     // se ci sono le mani, suona, altrimenti no, lo script oscillator osserva questa variabile per deicdere se deve suonare 
-    public static double[] current_Features;                //  attualmente le features sono floats, risolviamo sto problemo
+    public static float[] current_Features;                //  attualmente le features sono floats, risolviamo sto problemo
     public static int indexPlayingNote;                     //  indice della nota da suonare che è letta da PCMOscillator
     public static int indexPreviousNote;                    //  indice della nota suonata nel fixed update precedente
